Tests: different aspects in different files for better discovery
